Fix type of value assigned from strlen() in getline.c
This was declared as int, but strlen() in posix returns size_t.

https://pubs.opengroup.org/onlinepubs/009604499/functions/strlen.html

+++ src/native/getline.c
@@ -982,7 +982,7 @@ char *p;
 /* makes a copy of the string */
 {
     char *s = 0;
-    int   len = strlen(p);
+    size_t   len = strlen(p);
     char *nl = strchr(p, '\n');
 
     if (nl) {
